+++ packages/web-ui-registration/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @rocket.chat/web-ui-registration
 
+## 2.0.5
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@2.0.5
+
+## 2.0.4
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@2.0.4
+
+## 2.0.3
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@2.0.3
+
+## 2.0.2
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@2.0.2
+
+## 2.0.1
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@2.0.1
+
 ## 2.0.0
 
 ### Patch Changes